Refactor to pytorch_type_compatibility files
