+++ sdks/python/apache_beam/typehints/pytorch_type_compatibility_test.py
@@ -15,14 +15,10 @@
 # limitations under the License.
 #
 
-"""Unit tests for batched type converters."""
+"""Unit tests for pytorch_type_compabitility."""
 
-import contextlib
-import random
-import typing
 import unittest
 
-import numpy as np
 import pytest
 from parameterized import parameterized
 from parameterized import parameterized_class
@@ -30,34 +26,31 @@ from parameterized import parameterized_class
 from apache_beam.typehints import typehints
 from apache_beam.typehints.batch import BatchConverter
 from apache_beam.typehints.batch import N
-from apache_beam.typehints.batch import NumpyArray
+from apache_beam.typehints.pytorch_type_compatibility import PytorchTensor
+
+# Protect against environments where pytorch library is not available.
+# pylint: disable=wrong-import-order, wrong-import-position, ungrouped-imports
+try:
+  import torch
+except ImportError:
+  raise unittest.SkipTest('PyTorch dependencies are not installed')
 
 
-@pytest.mark.uses_pytorch
 @parameterized_class([
     {
-        'batch_typehint': np.ndarray,
-        'element_typehint': np.int32,
-        'batch': np.array(range(100), np.int32)
+        'batch_typehint': torch.Tensor,
+        'element_typehint': torch.Tensor,
+        'batch': torch.tensor(range(100), dtype=torch.int32)
     },
     {
-        'batch_typehint': NumpyArray[np.int64, (N, 10)],
-        'element_typehint': NumpyArray[np.int64, (10, )],
-        'batch': np.array([list(range(i, i + 10)) for i in range(100)],
-                          np.int64),
-    },
-    {
-        'batch_typehint': typehints.List[str],
-        'element_typehint': str,
-        'batch': ["foo" * (i % 5) + str(i) for i in range(1000)],
-    },
-    {
-        'batch_typehint': typing.List[str],
-        'element_typehint': str,
-        'batch': ["foo" * (i % 5) + str(i) for i in range(1000)],
+        'batch_typehint': PytorchTensor[torch.int64, (N, 10)],
+        'element_typehint': PytorchTensor[torch.int64, (10, )],
+        'batch': torch.tensor([list(range(i, i + 10)) for i in range(100)],
+                              dtype=torch.int64),
     },
 ])
-class BatchConverterTest(unittest.TestCase):
+@pytest.mark.uses_pytorch
+class DataFrameBatchConverterTest(unittest.TestCase):
   def create_batch_converter(self):
     return BatchConverter.from_typehints(
         element_type=self.element_typehint, batch_type=self.batch_typehint)
@@ -69,16 +62,16 @@ class BatchConverterTest(unittest.TestCase):
         self.element_typehint)
 
   def equality_check(self, left, right):
-    if isinstance(left, np.ndarray) and isinstance(right, np.ndarray):
-      return np.array_equal(left, right)
+    if isinstance(left, torch.Tensor):
+      self.assertTrue(torch.equal(left, right))
     else:
-      return left == right
+      raise TypeError(f"Encountered unexpected type, left is a {type(left)!r}")
 
   def test_typehint_validates(self):
     typehints.validate_composite_type_param(self.batch_typehint, '')
     typehints.validate_composite_type_param(self.element_typehint, '')
 
-  def test_type_check(self):
+  def test_type_check_batch(self):
     typehints.check_constraint(self.normalized_batch_typehint, self.batch)
 
   def test_type_check_element(self):
@@ -90,38 +83,24 @@ class BatchConverterTest(unittest.TestCase):
     rebatched = self.converter.produce_batch(exploded)
 
     typehints.check_constraint(self.normalized_batch_typehint, rebatched)
-    self.assertTrue(self.equality_check(self.batch, rebatched))
+    self.equality_check(self.batch, rebatched)
 
-  def test_estimate_byte_size_implemented(self):
-    # Just verify that we can call byte size
-    self.assertGreater(self.converter.estimate_byte_size(self.batch), 0)
-
-  @parameterized.expand([
-      (2, ),
-      (3, ),
-      (10, ),
-  ])
-  def test_estimate_byte_size_partitions(self, N):
+  def _split_batch_into_n_partitions(self, N):
     elements = list(self.converter.explode_batch(self.batch))
 
-    # Split elements into N contiguous partitions, create a batch out of each
-    batches = [
-        self.converter.produce_batch(
-            elements[len(elements) * i // N:len(elements) * (i + 1) // N])
+    # Split elements into N contiguous partitions
+    element_batches = [
+        elements[len(elements) * i // N:len(elements) * (i + 1) // N]
         for i in range(N)
     ]
 
-    # Some estimate_byte_size implementations use random samples,
-    # set a seed temporarily to make this test deterministic
-    with temp_seed(12345):
-      partitioned_size_estimate = sum(
-          self.converter.estimate_byte_size(batch) for batch in batches)
-      size_estimate = self.converter.estimate_byte_size(self.batch)
+    lengths = [len(element_batch) for element_batch in element_batches]
+    batches = [
+        self.converter.produce_batch(element_batch)
+        for element_batch in element_batches
+    ]
 
-    # Assert that size estimate for partitions is within 10% of size estimate
-    # for the whole partition.
-    self.assertLessEqual(
-        abs(partitioned_size_estimate / size_estimate - 1), 0.1)
+    return batches, lengths
 
   @parameterized.expand([
       (2, ),
@@ -129,19 +108,23 @@ class BatchConverterTest(unittest.TestCase):
       (10, ),
   ])
   def test_combine_batches(self, N):
-    elements = list(self.converter.explode_batch(self.batch))
-
-    # Split elements into N contiguous partitions, create a batch out of each
-    batches = [
-        self.converter.produce_batch(
-            elements[len(elements) * i // N:len(elements) * (i + 1) // N])
-        for i in range(N)
-    ]
+    batches, _ = self._split_batch_into_n_partitions(N)
 
     # Combine the batches, output should be equivalent to the original batch
     combined = self.converter.combine_batches(batches)
 
-    self.assertTrue(self.equality_check(self.batch, combined))
+    self.equality_check(self.batch, combined)
+
+  @parameterized.expand([
+      (2, ),
+      (3, ),
+      (10, ),
+  ])
+  def test_get_length(self, N):
+    batches, lengths = self._split_batch_into_n_partitions(N)
+
+    for batch, expected_length in zip(batches, lengths):
+      self.assertEqual(self.converter.get_length(batch), expected_length)
 
   def test_equals(self):
     self.assertTrue(self.converter == self.create_batch_converter())
@@ -149,17 +132,3 @@ class BatchConverterTest(unittest.TestCase):
 
   def test_hash(self):
     self.assertEqual(hash(self.create_batch_converter()), hash(self.converter))
-
-
-@contextlib.contextmanager
-def temp_seed(seed):
-  state = random.getstate()
-  random.seed(seed)
-  try:
-    yield
-  finally:
-    random.setstate(state)
-
-
-if __name__ == '__main__':
-  unittest.main()
